pyproject.toml
[tool.poetry]
name = "chess-ai-agent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
python-chess = "^1.999"
jinja2 = "^3.1.6"
openai = "^1.77.0"
python-dotenv = "^1.1.0"


[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
